added code for day 17 2022 part 1 and 2

+++ src/year2022/mod.rs
@@ -1,7 +1,7 @@
 use aoc::solution;
 use aoc::utils::solution::*;
 
-pub fn run(day: u8) -> Solution {
+pub fn run(day: u32) -> Solution {
     match day {
         1 => solution!(year2022, day1),
         2 => solution!(year2022, day2),
